Moved Textfield to legacy folder
