+++ packages/react/src/components/legacy/TextField/TextField.tsx
@@ -9,9 +9,9 @@ import type {
 } from 'react-number-format';
 import { NumericFormat, PatternFormat } from 'react-number-format';
 
-import { isNumericFormat, isPatternFormat } from '../../utils';
-import { InputWrapper } from '../_InputWrapper';
-import type { ReadOnlyVariant_, CharacterLimit } from '../_InputWrapper';
+import { isNumericFormat, isPatternFormat } from '../../../utils';
+import { InputWrapper } from '../../_InputWrapper';
+import type { ReadOnlyVariant_, CharacterLimit } from '../../_InputWrapper';
 
 export type TextFieldProps = {
   /**
